requirements.txt
scikit-image
scipy
opencv-python
tqdm
zarr
segmentation-models-pytorch
toml
numcodecs
timm==0.9.6
acvl_utils
pandas
dynamic_network_architectures
nibabel
matplotlib
seaborn
tifffile
Pillow
SimpleITK